requirements.txt
phiflow[jax]
jax
matplotlib
numpy
pyvista
pandas
scipy
